Completely removed gmenu2x...
Added a playlist generator for RetroArch to still get a nice pick up and play experience!

It scans for roms on boot and creates temporary playlists for each system.

It does slow down the boot but it seems acceptable even with 10,000 roms. Of course ideally we wouldn't rescan if nothing has changed but I have yet to find a reliable way of detecting that...